This fork is for the experimentation of the RTOS used in the rom-623
This fork tests the RTOS in different applications such as i a home security system
This fork reformatted the tasks/ and sensors/ directories.
These reformats are temporary and were done for visual reasons only.
The reason these must be seperate is that sensors/ act similar to device drivers as they are mutually inclusive to tasks meaning multiple tasks could use them to communicate with sensors
This leavses no good purose for them to be in the same directory except for me to meet guidlines of an institution who is reviewing this project